Update layout and style
