+++ Package.resolved
@@ -15,7 +15,7 @@
         "repositoryURL": "https://github.com/SwiftDocOrg/GraphViz.git",
         "state": {
           "branch": null,
-          "revision": "ce0327a62c4b5a5ddd4741406ea54f164f228a54",
+          "revision": "03405c13dc1c31f50c08bbec6e7587cbee1c7fb3",
           "version": null
         }
       },
@@ -42,8 +42,8 @@
         "repositoryURL": "https://github.com/apple/swift-argument-parser.git",
         "state": {
           "branch": null,
-          "revision": "35b76bf577d3cc74820f8991894ce3bcdf024ddc",
-          "version": "0.0.2"
+          "revision": "8d31a0905c346a45c87773ad50862b5b3df8dff6",
+          "version": "0.0.4"
         }
       },
       {
@@ -51,8 +51,8 @@
         "repositoryURL": "https://github.com/SwiftDocOrg/swift-cmark.git",
         "state": {
           "branch": null,
-          "revision": "2a766030bee955b4806044fd7aca1b6884475138",
-          "version": "0.28.3+20200110.2a76603"
+          "revision": "1168665f6b36be747ffe6b7b90bc54cfc17f42b7",
+          "version": "0.28.3+20200207.1168665"
         }
       },
       {
@@ -87,7 +87,7 @@
         "repositoryURL": "https://github.com/SwiftDocOrg/SwiftSemantics.git",
         "state": {
           "branch": "swift-5.2",
-          "revision": "3bb09e896a813cf0a72ec88a19dc0a7db49577cd",
+          "revision": "4fdc48bddbbb8311079ed111e5a4f2b92423b05c",
           "version": null
         }
       },
