+++ Sources/swift-doc/Extensions/SwiftDoc+Extensions.swift
@@ -8,6 +8,12 @@ extension Symbol {
     var node: Node {
         var node = Node(id.description)
         node.fontName = "Menlo"
+        node.shape = .box
+        node.style = .rounded
+
+        node.width = 3
+        node.height = 0.5
+        node.fixedSize = .shape
 
         if !(api is Unknown) {
             node.href = "/" + path(for: self)
@@ -15,16 +21,18 @@ extension Symbol {
 
         switch api {
         case let `class` as Class:
-            node.shape = .ellipse
+            node.class = "class"
             if `class`.modifiers.contains(where: { $0.name == "final" }) {
                 node.strokeWidth = 2.0
             }
+        case is Enumeration:
+            node.class = "enumeration"
         case is Structure:
-            node.shape = .box
-            node.style = .rounded
+            node.class = "structure"
         case is Protocol:
-            node.shape = .ellipse
-            node.style = .dashed
+            node.class = "protocol"
+        case is Unknown:
+            node.class = "unknown"
         default:
             break
         }
@@ -42,6 +50,7 @@ extension Symbol {
 
         var symbolNode = self.node
         symbolNode.strokeWidth = 3.0
+        symbolNode.class = [symbolNode.class, "current"].compactMap { $0 }.joined(separator: " ")
 
         graph.append(symbolNode)
 
@@ -64,12 +73,8 @@ extension Relationship {
         let to = object.node
 
         var edge = Edge(from: from.id, to: to.id)
-        switch predicate {
-        case .conformsTo:
-            edge.style = .dashed
-        default:
-            break
-        }
+        edge.class = predicate.rawValue
+        edge.preferredEdgeLength = 1.5
 
         return edge
     }
