+++ Sources/swift-doc/Supporting Types/Pages/TypePage.swift
@@ -20,27 +20,17 @@ struct TypePage: Page {
     }
 
     var document: CommonMark.Document {
-        return Document {
+        return CommonMark.Document {
             Heading { symbol.id.description }
 
             Documentation(for: symbol, in: module)
-
-            Inheritance(of: symbol, in: module)
-
-            if symbol.api is Protocol {
-                ConformingTypes(to: symbol, in: module)
-            } else if symbol.api is Type {
-                NestedTypes(of: symbol, in: module)
-            }
-
+            Relationships(of: symbol, in: module)
             Members(of: symbol, in: module)
             Requirements(of: symbol, in: module)
         }
     }
 
     var html: HypertextLiteral.HTML {
-//        print(symbol.api, module.interface.relationshipsBySubject[symbol.id])
-//        print("")
         return #"""
         <h1>
             <small>\#(String(describing: type(of: symbol.api)))</small>
@@ -48,7 +38,7 @@ struct TypePage: Page {
         </h1>
 
         \#(Documentation(for: symbol, in: module).html)
-        \#(Inheritance(of: symbol, in: module).html)
+        \#(Relationships(of: symbol, in: module).html)
         \#(Members(of: symbol, in: module).html)
         \#(Requirements(of: symbol, in: module).html)
         """#
